+++ R/bind.R
@@ -5,6 +5,8 @@
 #'
 #' @param ... data.tables or data.frames to bind
 #' @param .id If TRUE, an integer column is made as a group id
+#' @param use.names If TRUE, makes sure column names align
+#' @param fill If TRUE, fills missing columns with NA
 #'
 #' @export
 #' @md
@@ -22,12 +24,12 @@
 #'   bind_cols.(df2)
 #'
 #' bind_cols.(list(df1, df2))
-bind_rows. <- function(..., .id = NULL) {
+bind_rows. <- function(..., .id = NULL, use.names = TRUE, fill = TRUE) {
   UseMethod("bind_rows.")
 }
 
 #' @export
-bind_rows..default <- function(..., .id = NULL) {
+bind_rows..default <- function(..., .id = NULL, use.names = TRUE, fill = TRUE) {
 
   dots <- list(...)
   dots <- squash(dots)
@@ -35,7 +37,7 @@ bind_rows..default <- function(..., .id = NULL) {
   if (!all(dt_map_lgl(dots, is.data.table)))
     dots <- dt_map(dots, as_tidytable)
 
-  dots <- rbindlist(dots, idcol = .id)
+  dots <- rbindlist(dots, idcol = .id, use.names = use.names, fill = fill)
 
   as_tidytable(dots)
 }